Disable update check in default configuration
I'm unable to keep up with the updates that are implemented in the corecycler script. The only thing I've made is the GUI, so when the author of the corecycler script updates his script, it can cause issues with the GUI because the GUI is not updated to reflect the changes in the script. The version of Corecycler that the GUI was made to work with is included in this repository. I'm disabling the update checkbox as default so it doesn't cause issues.